requirements.txt
aiohttp>=3.8.5
aiosignal>=1.3.1
astunparse>=1.6.3
async-timeout>=4.0.3
asyncpg>=0.28.0
attrs>=23.1.0
braceexpand>=0.1.7
charset-normalizer>=3.2.0
click>=8.1.7
discord>=2.3.2
discord.py>=2.3.2
frozenlist>=1.4.0
idna>=3.4
import-expression>=1.1.4
jishaku>=2.5.1
line-profiler>=4.1.0
logzero>=1.7.0
multidict>=6.0.4
PyYAML>=6.0.1
six>=1.16.0
yarl>=1.9.2
